# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Codecademy e-Commerce REST API",
    "description": "Defines the backend logic relating to the 'registration', 'customers', 'addresses', 'orders', 'products', 'cart', and 'login' endpoints of an e-commerce REST API.",
    "version": "0.1"
  },
  "servers": [
    {
      "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1",
      "description": "SwaggerHub API Auto Mocking"
    },
    {
      "url": "https://codecademy-e-commerce-rest-api.herokuapp.com",
      "description": "Heroku App"
    }
  ],
  "paths": {
    "/api/registration": {
      "post": {
        "tags": [
          "Registration"
        ],
        "summary": "Register a new customer",
        "description": "Adds a new customer's details to the 'customers' and 'addresses' tables",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Registration"
              },
              "examples": {
                "0": {
                  "value": "{\n    \"first_name\": \"Customer First Name\",\n    \"last_name\": \"Customer Last Name\",\n    \"username\": \"Customer Username\",\n    \"password\": \"Customer Password\",\n    \"email\": \"Customer Email\",\n    \"street_number\": 7,\n    \"street_name\": \"Customer Street Name\",\n    \"town\": \"Customer Town\",\n    \"county\": \"Customer County\",\n    \"country\": \"Customer Country\",\n    \"postcode\": \"P05t c0d3\"\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns a confirmation of successful customer registration",
            "content": {
              "text/html; charset=utf-8": {
                "schema": {
                  "type": "string",
                  "example": "Your customer profile has been created successfully"
                }
              }
            }
          },
          "500": {
            "description": "Returns a description of the error that was unexpectedly encountered",
            "content": {
              "text/html; charset=utf-8": {
                "schema": {
                  "type": "string",
                  "example": "DEPENDING ON THE NATURE OF THE ERROR {err.message} THAT WAS ENCOUNTERED - The username you have selected is already in use. Please choose a different username. -OR- The password you have selected is already in use. Please choose a different password. - OR - The email address you have selected is already in use. Please choose a different email address. - OR - The unexpected value of {err.message}."
                }
              }
            }
          }
        },
        "servers": [
          {
            "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1"
          }
        ]
      },
      "servers": [
        {
          "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1"
        }
      ]
    },
    "/api/customers": {
      "get": {
        "tags": [
          "Customers"
        ],
        "summary": "Return all customer information",
        "description": "Returns the 'id', 'first_name', 'last_name', 'username', 'password', and 'email' fields for all customers in the 'customers' table",
        "responses": {
          "200": {
            "description": "JSON array of customers",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Customer"
                  }
                },
                "example": [
                  {
                    "id": 1,
                    "first_name": "Christopher",
                    "last_name": "Larham",
                    "username": "chris_larham_1983",
                    "password": "updated_chrissy_b0i*",
                    "email": "chrislarham@example.com"
                  },
                  {
                    "id": 2,
                    "first_name": "Charlotte",
                    "last_name": "Marie",
                    "username": "charlotte_marie_1986",
                    "password": "mar13p0os*",
                    "email": "charlottemarie@example.com"
                  },
                  {
                    "id": 3,
                    "first_name": "William",
                    "last_name": "Larham",
                    "username": "will_larham_2007",
                    "password": "w1lliam*",
                    "email": "williamlarham@example.com"
                  },
                  {
                    "id": 4,
                    "first_name": "Seth",
                    "last_name": "Larham",
                    "username": "seth_larham_2008",
                    "password": "s3th0*",
                    "email": "sethlarham@example.com"
                  },
                  {
                    "id": 5,
                    "first_name": "Alyssa",
                    "last_name": "Larham",
                    "username": "alyssa_larham_2010",
                    "password": "aly55a*",
                    "email": "alyssalarham@example.com"
                  },
                  {
                    "id": 6,
                    "first_name": "Roger",
                    "last_name": "Larham",
                    "username": "roger_larham_1955",
                    "password": "r0ger*",
                    "email": "rogerlarham@example.com"
                  },
                  {
                    "id": 7,
                    "first_name": "Judith",
                    "last_name": "Larham",
                    "username": "judi_larham_1956",
                    "password": "Jud1*",
                    "email": "judilarham@example.com"
                  },
                  {
                    "id": 8,
                    "first_name": "Matthew",
                    "last_name": "Rowlings",
                    "username": "matt_rowlings_1983",
                    "password": "Matti3*",
                    "email": "mattrowlings@example.com"
                  },
                  {
                    "id": 9,
                    "first_name": "Steen",
                    "last_name": "Jinks",
                    "username": "steen_jinks_1991",
                    "password": "Jink5*",
                    "email": "steenjinks@example.com"
                  },
                  {
                    "id": 10,
                    "first_name": "Joe",
                    "last_name": "Bloggs",
                    "username": "joe_bloggs_1992",
                    "password": "bl0ggs",
                    "email": "joebloggs@example.com"
                  },
                  {
                    "id": 11,
                    "first_name": "Lugali",
                    "last_name": "Bobtail",
                    "username": "lugali_bobtail_1758",
                    "password": "lugal1*",
                    "email": "lugalibobtail@example.com"
                  },
                  {
                    "id": 12,
                    "first_name": "Divine",
                    "last_name": "Intervention",
                    "username": "divine_intervention_0",
                    "password": "Genes1s*",
                    "email": "divineintervention@outlook.com"
                  },
                  {
                    "id": 13,
                    "first_name": "Customer First Name",
                    "last_name": "Customer Last Name",
                    "username": "Customer Username",
                    "password": "Customer Password",
                    "email": "Customer Email"
                  }
                ]
              }
            }
          },
          "500": {
            "description": "Returns a description of the error {err} that was unexpectedly encountered",
            "content": {
              "text/html; charset=utf-8": {
                "schema": {
                  "type": "string",
                  "example": "The unexpected value of {err.message}."
                }
              }
            }
          }
        },
        "servers": [
          {
            "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1"
          },
          {
            "url": "https://codecademy-e-commerce-rest-api.herokuapp.com",
            "description": "Heroku App"
          }
        ]
      },
      "servers": [
        {
          "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1"
        }
      ]
    },
...
    "/api/addresses": {
      "get": {
        "tags": [
          "Addresses"
        ],
        "summary": "Return the addresses for all customers",
        "description": "Returns all the addresses in the 'addresses' table",
        "responses": {
          "200": {
            "description": "Returns a JSON array of addresses",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Address"
                  }
                },
                "example": [
                  {
                    "id": 1,
                    "street_number": 18,
                    "town": "St. Austell",
                    "county": "Cornwall",
                    "country": "England",
                    "postcode": "PL26 8TS",
                    "customer_id": 1,
                    "street_name": "Rosina Way"
                  },
                  {
                    "id": 2,
                    "street_number": 18,
                    "town": "St. Austell",
                    "county": "Cornwall",
                    "country": "England",
                    "postcode": "PL26 8TS",
                    "customer_id": 2,
                    "street_name": "Rosina Way"
                  },
                  {
                    "id": 3,
                    "street_number": 18,
                    "town": "St. Austell",
                    "county": "Cornwall",
                    "country": "England",
                    "postcode": "PL26 8TS",
                    "customer_id": 3,
                    "street_name": "Rosina Way"
                  },
                  {
                    "id": 4,
                    "street_number": 18,
                    "town": "St. Austell",
                    "county": "Cornwall",
                    "country": "England",
                    "postcode": "PL26 8TS",
                    "customer_id": 4,
                    "street_name": "Rosina Way"
                  },
                  {
                    "id": 5,
                    "street_number": 18,
                    "town": "St. Austell",
                    "county": "Cornwall",
                    "country": "England",
                    "postcode": "PL26 8TS",
                    "customer_id": 5,
                    "street_name": "Rosina Way"
                  },
                  {
                    "id": 6,
                    "street_number": 31,
                    "town": "Bugle",
                    "county": "Cornwall",
                    "country": "England",
                    "postcode": "PL26 8SJ",
                    "customer_id": 10,
                    "street_name": "Springfields"
                  },
                  {
                    "id": 7,
                    "street_number": 125,
                    "town": "Upville",
                    "county": "Super Mario",
                    "country": "Oogloo",
                    "postcode": "P249 IOLUG",
                    "customer_id": 11,
                    "street_name": "Rainbow Hills"
                  },
                  {
                    "id": 8,
                    "street_number": 7,
                    "town": "Eden",
                    "county": "Paradise",
                    "country": "Omnipresence",
                    "postcode": "EVR1 WH3R3",
                    "customer_id": 12,
                    "street_name": "Day Creation"
                  },
                  {
                    "id": 9,
                    "street_number": 9,
                    "town": "St. Austell",
                    "county": "Cornwall",
                    "country": "England",
                    "postcode": "PL25 4BJ",
                    "customer_id": 6,
                    "street_name": "North Hill Park"
                  },
                  {
                    "id": 10,
                    "street_number": 9,
                    "town": "St. Austell",
                    "county": "Cornwall",
                    "country": "England",
                    "postcode": "PL25 4BJ",
                    "customer_id": 7,
                    "street_name": "North Hill Park"
                  },
                  {
                    "id": 11,
                    "street_number": 7,
                    "town": "St. Day",
                    "county": "Cornwall",
                    "country": "England",
                    "postcode": "TR16 5MR",
                    "customer_id": 8,
                    "street_name": "Trenance Road"
                  },
                  {
                    "id": 12,
                    "street_number": 10,
                    "town": "Redruth",
                    "county": "Cornwall",
                    "country": "England",
                    "postcode": "TR15 1SJ",
                    "customer_id": 9,
                    "street_name": "Random Road"
                  },
                  {
                    "id": 13,
                    "street_number": 14,
                    "town": "Updated Town",
                    "county": "Updated County",
                    "country": "Updated Country",
                    "postcode": "P05t c0d3",
                    "customer_id": 13,
                    "street_name": "Updated Street Name"
                  }
                ]
              }
            }
          },
          "500": {
            "description": "Returns a description of the error {err} that was unexpectedly encountered",
            "content": {
              "text/html; charset=utf-8": {
                "schema": {
                  "type": "string",
                  "example": "The unexpected value of {err.message}."
                }
              }
            }
          }
        },
        "servers": [
          {
            "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1"
          }
        ]
      },
      "servers": [
        {
          "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1"
        }
      ]
    },
...
    "/api/products": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Return all product information",
        "description": "Returns an array of JSON objects that describe the products in the database",
        "responses": {
          "200": {
            "description": "Indicates the successful return of an array of JSON objects that describe the products in the database",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                },
                "example": [
                  {
                    "id": 1,
                    "item_name": "Opportunity Rover - 2008-04-03 (Sol 1490)",
                    "item_description": "Photo taken by the Opportunity Rover's Navigation Camera on 2008-04-03 (expedition sol 1490)",
                    "image_url": "https//mars.nasa.gov/mer/gallery/all/1/n/1490/1N260454912EFF8976P1947L0M1-BR.JPG",
                    "price": 1000
                  },
                  {
                    "id": 2,
                    "item_name": "Opportunity Rover - 2004-02-04 (Sol 10)",
                    "item_description": "Photo taken by the Opportunity Rover's Front Hazard Avoidance Camera on 2004-02-04 (expedition sol 10)",
                    "image_url": "https://mars.nasa.gov/mer/gallery/all/1/f/010/1F129070908EDN0224P1101L0M1-BR.JPG",
                    "price": 800
                  },
                  {
                    "id": 3,
                    "item_name": "Opportunity Rover - 2004-01-26 (Sol 1)",
                    "item_description": "Photo taken by the Opportunity Rover's Panoramic Camera on 2004-01-26 (expedition sol 1)",
                    "image_url": "https://mars.nasa.gov/mer/gallery/all/1/p/001/1P128287426EFF0000P2303L2M1-BR.JPG",
                    "price": 600
                  },
                  {
                    "id": 4,
                    "item_name": "Opportunity Rover - 2009-09-11 (Sol 2001)",
                    "item_description": "Photo taken by the Opportunity Rover's Rear Hazard   Avoidance Camera on 2009-09-11 (expedition sol 2001)",
                    "image_url": "https://mars.nasa.gov/mer/gallery/all/1/r/2001/1R305828935EFFA5FWP1312L0M1-BR.JPG",
                    "price": 700
                  },
                  {
                    "id": 5,
                    "item_name": "Opportunity Rover - 2004-01-26 (Sol 1)",
                    "item_description": "Photo taken by the Opportunity Rover's Rear Hazard Avoidance Camera on 2004-01-26 (expedition sol 1)",
                    "image_url": "https://mars.nasa.gov/mer/gallery/all/1/r/001/1R128285184EDN0000P1002L0M1-BR.JPG",
                    "price": 900
                  },
                  {
                    "id": 6,
                    "item_name": "Opportunity Rover - 2006-11-18 (Sol 1001)",
                    "item_description": "Photo taken by the Opportunity Rover's Panoramic Camera on 2006-11-18 (expedition sol 1001)",
                    "image_url": "https://mars.nasa.gov/mer/gallery/all/1/p/1001/1P217047759EFF76ZFP2426L2M1-BR.JPG",
                    "price": 1500
                  },
                  {
                    "id": 7,
                    "item_name": "Opportunity Rover - 2009-09-11 (Sol 2001)",
                    "item_description": "Photo taken by the Opportunity Rover's Front Hazard Avoidance Camera on 2009-09-11 (expedition sol 2001)",
                    "image_url": "https://mars.nasa.gov/mer/gallery/all/1/f/2001/1F305828985EFFA5FWP1212L0M1-BR.JPG",
                    "price": 1250
                  },
                  {
                    "id": 8,
                    "item_name": "Item Name [varchar(50)]",
                    "item_description": "Item Description [text]",
                    "image_url": "The NEW URL of the image associated with the item [varchar(100)]",
                    "price": 1000
                  }
                ]
              }
            }
          },
          "500": {
            "description": "Returns a description of the error {err} that was unexpectedly encountered",
            "content": {
              "text/html; charset=utf-8": {
                "schema": {
                  "type": "string",
                  "example": "The unexpected value of {err.message}."
                }
              }
            }
          }
        },
        "servers": [
          {
            "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1"
          }
        ]
      },
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Add a product to the database",
        "description": "Adds a product to the 'products' table of the e-Commerce database",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreation"
              },
              "examples": {
                "0": {
                  "value": "{\n    \"item_name\": \"Item Name [varchar(50)]\",\n    \"item_description\": \"Item Description [text]\",\n    \"image_url\": \"The URL of the image associated with the item [varchar(100)]\",\n    \"price\": 1000\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "String response that confirms product creation success",
            "content": {
              "text/html; charset=utf-8": {
                "schema": {
                  "type": "string"
                },
                "example": "Product created successfully."
              }
            }
          },
          "409": {
            "description": "The product's 'image_url' is already in use and thus violates the unique constraint",
            "content": {
              "text/html; charset=utf-8": {
                "schema": {
                  "type": "string",
                  "example": "Product creation did not succeed because the image URL is already in use."
                }
              }
            }
          },
          "500": {
            "description": "Returns a description of the error {err} that was unexpectedly encountered",
            "content": {
              "text/html; charset=utf-8": {
                "schema": {
                  "type": "string",
                  "example": "The unexpected value of {err.message}."
                }
              }
            }
          }
        },
        "servers": [
          {
            "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1"
          }
        ]
      },
      "servers": [
        {
          "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1"
        }
      ]
    },
...
  },
  "components": {
    "schemas": {
      "Registration": {
        "type": "object",
        "properties": {
          "first_name": {
            "type": "string",
            "description": "The customer's first name"
          },
          "last_name": {
            "type": "string",
            "description": "The customer's last name"
          },
          "username": {
            "type": "string",
            "description": "The customer's username"
          },
          "password": {
            "type": "string",
            "description": "The customer's password"
          },
          "email": {
            "type": "string",
            "description": "The customer's email address"
          },
          "street_number": {
            "type": "integer",
            "description": "The street number associated with the customer's address"
          },
          "street_name": {
            "type": "string",
            "description": "The street name associated with the customer's address"
          },
          "town": {
            "type": "string",
            "description": "The town associated with the customer's address"
          },
          "county": {
            "type": "string",
            "description": "The county associated with the customer's address"
          },
          "country": {
            "type": "string",
            "description": "The country associated with the customer's address"
          },
          "postcode": {
            "type": "string",
            "description": "The postcode associated with the customer's address"
          }
        }
      },
      "Customer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "The customer's ID"
          },
          "first_name": {
            "type": "string",
            "description": "The customer's first name"
          },
          "last_name": {
            "type": "string",
            "description": "The customer's last name"
          },
          "username": {
            "type": "string",
            "description": "The customer's username"
          },
          "password": {
            "type": "string",
            "description": "The customer's password"
          },
          "email": {
            "type": "string",
            "description": "The customer's email address"
          }
        }
      },
...
      "Address": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "The address id"
          },
          "street_number": {
            "type": "integer",
            "description": "The street number"
          },
          "street_name": {
            "type": "string",
            "description": "The street name"
          },
          "town": {
            "type": "string",
            "description": "The town"
          },
          "county": {
            "type": "string",
            "description": "The county"
          },
          "country": {
            "type": "string",
            "description": "The country"
          },
          "postcode": {
            "type": "string",
            "description": "The postcode"
          },
          "customer_id": {
            "type": "integer",
            "description": "The id of the customer who lives at this address"
          }
        }
      },
...
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "The product ID"
          },
          "item_name": {
            "type": "string",
            "description": "The product name"
          },
          "item_description": {
            "type": "string",
            "description": "The product description"
          },
          "image_url": {
            "type": "string",
            "description": "The URL of the image that represents the item"
          },
          "price": {
            "type": "integer",
            "description": "The price of the product in pence"
          }
        }
      },
      "ProductCreation": {
        "type": "object",
        "properties": {
          "item_name": {
            "type": "string",
            "description": "The product name"
          },
          "item_description": {
            "type": "string",
            "description": "The product description"
          },
          "image_url": {
            "type": "string",
            "description": "The URL of the image that represents the item"
          },
          "price": {
            "type": "integer",
            "description": "The price of the product in pence"
          }
        }
      },
...
    }
  }
}